Project restructure (os & random)
