+++ engine/benchmark/CMakeLists.txt
@@ -16,6 +16,7 @@ function(add_engine_benchmark subdirectory)
 
   add_executable(${subdirectory}_benchmark ${${subdirectory}_bench_headers} ${${subdirectory}_bench_sources})
 
+  target_include_directories(${subdirectory}_benchmark PRIVATE ${CMAKE_CURRENT_SOURCE_DIR})
   target_include_directories(${subdirectory}_benchmark PRIVATE ${engine_include_directories} ${benchmark_INCLUDE_DIRS})
   target_link_libraries(${subdirectory}_benchmark PRIVATE engine ${benchmark_LIBRARIES})
 endfunction()
@@ -28,4 +29,5 @@ add_engine_benchmark(util)
 add_engine_benchmark(events)
 add_engine_benchmark(graphics)
 add_engine_benchmark(ecs)
-add_engine_benchmark(parallel)
+add_engine_benchmark(async)
+add_engine_benchmark(random)
