+++ engine/test/CMakeLists.txt
@@ -31,4 +31,6 @@ add_engine_test(util)
 add_engine_test(events)
 add_engine_test(graphics)
 add_engine_test(ecs)
-add_engine_test(parallel)
+add_engine_test(async)
+add_engine_test(os)
+add_engine_test(random)
